Move `ButtonVisuals` to their own file
